reorganisation de l'api et ajustement des requetes en frontend
