+++ backend/db/index.js
@@ -1,7 +1,6 @@
 const { Pool } = require('pg');
 require('dotenv/config');
 
-
 const pool = new Pool({
     connectionString: process.env.DB_CONNECTION,
 });
