+++ backend/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon server/index.js",
+    "start": "nodemon app.js",
     "frontend": "cd ../frontend && npm run start",
-    "dev": "concurrently \"npm run start\" \"npm run frontend\""
+    "dev": "concurrently --kill-others-on-fail \"npm run start\" \"npm run frontend\""
   },
   "author": "",
   "license": "ISC",
