--- backend/server/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "express-fileupload": "^1.2.0",
-    "moment": "^2.29.0",
-    "morgan": "^1.10.0",
-    "nodemon": "^2.0.4",
-    "pg": "^8.3.3"
-  }
-}
