+++ frontend/package.json
@@ -2,6 +2,7 @@
   "name": "projet",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:5000",
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
@@ -30,7 +31,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "backend": "cd ../backend && npm run start",
-    "dev": "concurrently \"npm run backend\" \"npm run start\""
+    "dev": "concurrently --kill-others-on-fail \"npm run backend\" \"npm run start\""
   },
   "eslintConfig": {
     "extends": "react-app"
