+++ frontend/src/components/Home/Home.js
@@ -10,7 +10,7 @@ export function Home() {
 
   const getProjects = async () => {
     try {
-      const response = await fetch("http://localhost:5000/Accueil");
+      const response = await fetch("/project/top3");
       const jsonData = await response.json();
 
       setProjects(jsonData);
